// maplibre/src/client/models/LinesGeoJson.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { GeoFeature_LineStringGeometry_ } from './GeoFeature_LineStringGeometry_';

export type LinesGeoJson = {
  type?: string;
  features: Array<GeoFeature_LineStringGeometry_>;
};