// maplibre/src/client/models/PointsGeoJson.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { GeoFeature_PointGeometry_ } from './GeoFeature_PointGeometry_';

export type PointsGeoJson = {
  type?: string;
  features: Array<GeoFeature_PointGeometry_>;
};